Moving recoverFromFiles to Translog class
